requirements.txt
requests>=2.31.0
piper-tts>=0.1.0
openai-whisper @ git+https://github.com/openai/whisper.git
python-dotenv>=1.0.0

